@@ -11,36 +11,37 @@
   },
   "dependencies": {
     "@quasar/extras": "1.16.11",
-    "apexcharts": "3.48.0",
-    "axios": "1.6.8",
+    "@vueuse/core": "10.11.0",
+    "@vueuse/integrations": "10.11.0",
+    "@vueuse/shared": "10.11.0",
+    "apexcharts": "3.49.2",
+    "axios": "1.7.2",
     "dotenv": "16.4.5",
-    "pinia": "^2.1.7",
-    "qrcode.vue": "3.4.1",
-    "quasar": "2.15.3",
-    "vue": "3.4.23",
-    "vue3-apexcharts": "1.5.2",
+    "monaco-editor": "0.50.0",
+    "pinia": "2.1.7",
+    "qrcode": "1.5.3",
+    "quasar": "2.16.4",
+    "vue": "3.4.31",
+    "vue-router": "4.4.0",
+    "vue3-apexcharts": "1.5.3",
     "vuedraggable": "4.1.0",
-    "vue-router": "4.3.2",
-    "@vueuse/core": "10.9.0",
-    "@vueuse/shared": "10.9.0",
-    "monaco-editor": "0.47.0",
     "vuex": "4.1.0",
-    "xterm": "^5.3.0",
-    "xterm-addon-fit": "^0.8.0",
-    "yaml": "2.4.1"
+    "@xterm/xterm": "5.5.0",
+    "@xterm/addon-fit": "0.10.0",
+    "yaml": "2.4.5"
   },
   "devDependencies": {
-    "@quasar/cli": "2.4.0",
     "@intlify/unplugin-vue-i18n": "4.0.0",
-    "@quasar/app-vite": "1.8.1",
-    "@types/node": "20.12.7",
-    "@typescript-eslint/eslint-plugin": "7.7.0",
-    "@typescript-eslint/parser": "7.7.0",
+    "@quasar/app-vite": "1.9.3",
+    "@quasar/cli": "2.4.1",
+    "@types/node": "20.14.9",
+    "@typescript-eslint/eslint-plugin": "7.14.1",
+    "@typescript-eslint/parser": "7.14.1",
     "autoprefixer": "10.4.19",
     "eslint": "8.57.0",
     "eslint-config-prettier": "9.1.0",
     "eslint-plugin-vue": "8.7.1",
     "prettier": "3.2.5",
-    "typescript": "5.4.5"
+    "typescript": "5.5.2"
   }
 }
