@@ -36,7 +36,7 @@
 <script>
 // composition imports
 import { ref, onMounted } from "vue";
-import { testScript } from "@/api/scripts";
+import { testScript, testScriptOnServer } from "@/api/scripts";
 import { useDialogPluginComponent } from "quasar";
 
 export default {
@@ -45,6 +45,7 @@ export default {
   props: {
     script: !Object,
     agent: !String,
+    ctx: !String,
   },
   setup(props) {
     // setup quasar dialog plugin
@@ -70,7 +71,11 @@ export default {
         env_vars: props.script.env_vars,
       };
       try {
-        ret.value = await testScript(props.agent, data);
+        if (props.ctx === "server") {
+          ret.value = await testScriptOnServer(data);
+        } else {
+          ret.value = await testScript(props.agent, data);
+        }
       } catch (e) {
         console.error(e);
       }
