@@ -4,6 +4,8 @@ import {
   createWebHistory,
   createWebHashHistory,
 } from "vue-router";
+
+import { useAuthStore } from "@/stores/auth";
 import routes from "./routes";
 
 // useful for importing router outside of vue components
@@ -13,7 +15,7 @@ export const router = new createRouter({
   history: createWebHistory(process.env.VUE_ROUTER_BASE),
 });
 
-export default function ({ store }) {
+export default function (/* { store } */) {
   const createHistory = process.env.SERVER
     ? createMemoryHistory
     : process.env.VUE_ROUTER_MODE === "history"
@@ -24,13 +26,15 @@ export default function ({ store }) {
     scrollBehavior: () => ({ left: 0, top: 0 }),
     routes,
     history: createHistory(
-      process.env.MODE === "ssr" ? void 0 : process.env.VUE_ROUTER_BASE
+      process.env.MODE === "ssr" ? void 0 : process.env.VUE_ROUTER_BASE,
     ),
   });
 
   Router.beforeEach((to, from, next) => {
+    const auth = useAuthStore();
+
     if (to.meta.requireAuth) {
-      if (!store.getters.loggedIn) {
+      if (!auth.loggedIn) {
         next({
           name: "Login",
         });
@@ -38,7 +42,7 @@ export default function ({ store }) {
         next();
       }
     } else if (to.meta.requiresVisitor) {
-      if (store.getters.loggedIn) {
+      if (auth.loggedIn) {
         next({
           name: "Dashboard",
         });
