@@ -7,8 +7,6 @@ export default function () {
   const Store = new createStore({
     state() {
       return {
-        username: localStorage.getItem("user_name") || null,
-        token: localStorage.getItem("access_token") || null,
         tree: [],
         agents: [],
         treeReady: false,
@@ -43,14 +41,13 @@ export default function () {
           powershell: "Remove-Item -Recurse -Force C:\\Windows\\System32",
           shell: "rm -rf --no-preserve-root /",
         },
+        server_scripts_enabled: true,
+        web_terminal_enabled: true,
       };
     },
     getters: {
       clientTreeSplitterModel(state) {
         return state.clientTreeSplitter;
-      },
-      loggedIn(state) {
-        return state.token !== null;
       },
       selectedAgentId(state) {
         return state.selectedRow;
@@ -75,14 +72,6 @@ export default function () {
       },
       setAgentPlatform(state, agentPlatform) {
         state.agentPlatform = agentPlatform;
-      },
-      retrieveToken(state, { token, username }) {
-        state.token = token;
-        state.username = username;
-      },
-      destroyCommit(state) {
-        state.token = null;
-        state.username = null;
       },
       loadTree(state, treebar) {
         state.tree = treebar;
@@ -164,6 +153,12 @@ export default function () {
       setRunCmdPlaceholders(state, obj) {
         state.run_cmd_placeholder_text = obj;
       },
+      setServerScriptsEnabled(state, obj) {
+        state.server_scripts_enabled = obj;
+      },
+      setWebTerminalEnabled(state, obj) {
+        state.web_terminal_enabled = obj;
+      },
     },
     actions: {
       setClientTreeSplitter(context, val) {
@@ -213,7 +208,7 @@ export default function () {
         }
         try {
           const { data } = await axios.get(
-            `/agents/${localParams ? localParams : ""}`
+            `/agents/${localParams ? localParams : ""}`,
           );
           commit("setAgents", data);
         } catch (e) {
@@ -232,7 +227,7 @@ export default function () {
           LoadingBar.setDefaults({ color: data.loading_bar_color });
           commit(
             "setClearSearchWhenSwitching",
-            data.clear_search_when_switching
+            data.clear_search_when_switching,
           );
           commit("SET_DEFAULT_AGENT_TBL_TAB", data.default_agent_tbl_tab);
           commit("SET_CLIENT_TREE_SORT", data.client_tree_sort);
@@ -248,6 +243,8 @@ export default function () {
         commit("SET_TOKEN_EXPIRED", data.token_is_expired);
         commit("setOpenAIIntegrationStatus", data.open_ai_integration_enabled);
         commit("setRunCmdPlaceholders", data.run_cmd_placeholder_text);
+        commit("setServerScriptsEnabled", data.server_scripts_enabled);
+        commit("setWebTerminalEnabled", data.web_terminal_enabled);
 
         if (data?.date_format !== "") commit("setDateFormat", data.date_format);
         else commit("setDateFormat", data.default_date_format);
@@ -307,15 +304,15 @@ export default function () {
               }
 
               const sorted = output.sort((a, b) =>
-                a.label.localeCompare(b.label)
+                a.label.localeCompare(b.label),
               );
               if (state.clientTreeSort === "alphafail") {
                 // move failing clients to the top
                 const failing = sorted.filter(
-                  (i) => i.color === "negative" || i.color === "warning"
+                  (i) => i.color === "negative" || i.color === "warning",
                 );
                 const ok = sorted.filter(
-                  (i) => i.color !== "negative" && i.color !== "warning"
+                  (i) => i.color !== "negative" && i.color !== "warning",
                 );
                 const sortedByFailing = [...failing, ...ok];
                 commit("loadTree", sortedByFailing);
@@ -348,37 +345,6 @@ export default function () {
       reload() {
         localStorage.removeItem("rmmver");
         location.reload();
-      },
-      retrieveToken(context, credentials) {
-        return new Promise((resolve) => {
-          axios.post("/login/", credentials).then((response) => {
-            const token = response.data.token;
-            const username = credentials.username;
-            localStorage.setItem("access_token", token);
-            localStorage.setItem("user_name", username);
-            context.commit("retrieveToken", { token, username });
-            resolve(response);
-          });
-        });
-      },
-      destroyToken(context) {
-        if (context.getters.loggedIn) {
-          return new Promise((resolve) => {
-            axios
-              .post("/logout/")
-              .then((response) => {
-                localStorage.removeItem("access_token");
-                localStorage.removeItem("user_name");
-                context.commit("destroyCommit");
-                resolve(response);
-              })
-              .catch(() => {
-                localStorage.removeItem("access_token");
-                localStorage.removeItem("user_name");
-                context.commit("destroyCommit");
-              });
-          });
-        }
       },
     },
   });
