@@ -1,6 +1,10 @@
 import { Notify } from "quasar";
 
-export function isValidThreshold(warning, error, diskcheck = false) {
+export function isValidThreshold(
+  warning: number,
+  error: number,
+  diskcheck = false,
+) {
   if (warning === 0 && error === 0) {
     Notify.create({
       type: "negative",
@@ -31,7 +35,7 @@ export function isValidThreshold(warning, error, diskcheck = false) {
   return true;
 }
 
-export function validateEventID(val) {
+export function validateEventID(val: number | "*") {
   if (val === null || val.toString().replace(/\s/g, "") === "") {
     return false;
   } else if (val === "*") {
@@ -44,10 +48,20 @@ export function validateEventID(val) {
 }
 
 // validate script return code
-export function validateRetcode(val, done) {
+// function is used for quasar's q-select on-new-value function
+export function validateRetcode(
+  val: string,
+  done: (item?: unknown, mode?: "add" | "add-unique" | "toggle") => void,
+) {
   /^\d+$/.test(val) ? done(val) : done();
 }
 
-export function validateTimePeriod(val) {
+export function validateTimePeriod(val: string) {
   return /^\d{1,3}(H|h|M|m|S|s|d|D)$/.test(val);
+}
+
+export function isValidEmail(val: string) {
+  const email =
+    /^(?=[a-zA-Z0-9@._%+-]{6,254}$)[a-zA-Z0-9._%+-]{1,64}@(?:[a-zA-Z0-9-]{1,63}\.){1,8}[a-zA-Z]{2,63}$/;
+  return email.test(val);
 }
