@@ -824,7 +824,9 @@ export default {
           );
           return;
         }
-        this.urlActions = r.data;
+        this.urlActions = r.data.filter(
+          (action) => action.action_type === "web",
+        );
       });
     },
     runURLAction(id, action, model) {
